backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kushal Mamillapalli", email = "km6238@nyu.edu" }
]
requires-python = ">=3.13"
dependencies = [
    "eth-account>=0.13.7",
    "flasgger>=0.9.7.1",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "google-generativeai>=0.8.0",
    "openai>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "pymongo>=4.6.0",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "tavily-python>=0.3.0",
    "web3>=7.14.1",
]

[build-system]
requires = ["uv_build>=0.8.6,<0.9.0"]
build-backend = "uv_build"
